refactor: embed agent.py
