+++ cmd/ax/harness.go
@@ -30,9 +30,8 @@ import (
 )
 
 var (
-	harnessPort                 int
-	harnessHost                 string
-	harnessAntigravityAgentFile string
+	harnessPort int
+	harnessHost string
 )
 
 var harnessCmd = &cobra.Command{
@@ -45,7 +44,6 @@ var harnessCmd = &cobra.Command{
 func init() {
 	harnessCmd.Flags().IntVar(&harnessPort, "port", 50053, "Port for the HarnessService to listen on")
 	harnessCmd.Flags().StringVar(&harnessHost, "host", "127.0.0.1", "Host interface for the HarnessService to bind")
-	harnessCmd.Flags().StringVar(&harnessAntigravityAgentFile, "antigravity-agent-file", "python/antigravity/agent.py", "Path to the agent config file the Python sidecar serves")
 	rootCmd.AddCommand(harnessCmd)
 }
 
@@ -75,7 +73,6 @@ func runHarness(cmd *cobra.Command, args []string) error {
 	py := exec.Command("python3", "-m", "python.antigravity.harness_server",
 		"--host", harnessHost,
 		"--port", strconv.Itoa(harnessPort),
-		"--agent_file", harnessAntigravityAgentFile,
 	)
 	py.Stdin = os.Stdin
 	py.Stdout = os.Stdout
